// web-application/src/types-openapi/api/models/ListBlogRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListBlogRequest
 */
export interface ListBlogRequest {
    /**
     * 
     * @type {string}
     * @memberof ListBlogRequest
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the ListBlogRequest interface.
 */
export function instanceOfListBlogRequest(value: object): value is ListBlogRequest {
    return true;
}

export function ListBlogRequestFromJSON(json: any): ListBlogRequest {
    return ListBlogRequestFromJSONTyped(json, false);
}

export function ListBlogRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListBlogRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListBlogRequestToJSON(json: any): ListBlogRequest {
    return ListBlogRequestToJSONTyped(json, false);
}

export function ListBlogRequestToJSONTyped(value?: ListBlogRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'status': value['status'],
    };
}