// web-application/src/types-openapi/api/models/ListBlogResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListBlogResponse
 */
export interface ListBlogResponse {
    /**
     * 
     * @type {Date}
     * @memberof ListBlogResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ListBlogResponse
     */
    updatedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    createdBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    updatedBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    content?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListBlogResponse
     */
    status?: string | null;
}

/**
 * Check if a given object implements the ListBlogResponse interface.
 */
export function instanceOfListBlogResponse(value: object): value is ListBlogResponse {
    return true;
}

export function ListBlogResponseFromJSON(json: any): ListBlogResponse {
    return ListBlogResponseFromJSONTyped(json, false);
}

export function ListBlogResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListBlogResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'createdBy': json['createdBy'] == null ? undefined : json['createdBy'],
        'updatedBy': json['updatedBy'] == null ? undefined : json['updatedBy'],
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'content': json['content'] == null ? undefined : json['content'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListBlogResponseToJSON(json: any): ListBlogResponse {
    return ListBlogResponseToJSONTyped(json, false);
}

export function ListBlogResponseToJSONTyped(value?: ListBlogResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] === null ? null : ((value['updatedAt'] as any)?.toISOString()),
        'createdBy': value['createdBy'],
        'updatedBy': value['updatedBy'],
        'id': value['id'],
        'title': value['title'],
        'content': value['content'],
        'imageUrl': value['imageUrl'],
        'status': value['status'],
    };
}